requirements.txt
fastapi 
uvicorn 
psycopg2-binary 
requests 
pydantic